+++ NEWS.md
@@ -8,6 +8,8 @@
   number of tests performed, instead of the number of inputs tested.
 - Add the `tabulate` argument to `summary()` to control whether the tabulation
   of results should be printed to the terminal or not.
+- Change the internal structure of the `cbtf` object to store the names of
+  the functions fuzzed only once.
 
 # CBTF 0.4.0 (2025-07-31)